requirements.txt
# Data manipulation
pandas==2.0.3
numpy==1.24.3

# Machine Learning
scikit-learn==1.3.0
scipy==1.11.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Jupyter
jupyter==1.0.0
notebook==6.5.4
ipykernel==6.25.1
ipywidgets==8.1.0

# Utilities
python-dateutil==2.8.2
tqdm==4.65.0
pyyaml==6.0
python-dotenv==1.0.0

# Data download
requests==2.31.0